v2.3
Correción e eliminación de la clase menu
