+++ proyecto_programacion/E_S.java
@@ -5,26 +5,26 @@ class E_S {
 	
 	Scanner teclado= new Scanner(System.in);
 	
-	public String CadenaTexto (String dato) {
-		dato= teclado.nextLine();
+	public String CadenaTexto () {
+		String dato= teclado.nextLine();
 		
 		return dato;
 	}
 	
-	public int EnteroTexto (int dato) {
-		dato= teclado.nextInt();
+	public int EnteroTexto () {
+		int dato= teclado.nextInt();
 		
 		return dato;
 	}
 	
-	public double DoubleTexto (double dato) {
-		dato= teclado.nextDouble();
+	public double DoubleTexto () {
+		double dato= teclado.nextDouble();
 		
 		return dato;
 	}
 	
-	public char CaracterTexto (char dato) {
-		dato= teclado.next().charAt(0);
+	public char CaracterTexto () {
+		char dato= teclado.next().charAt(0);
 		
 		return dato;
 	}
