+++ proyecto_programacion/administracion.java
@@ -3,11 +3,354 @@ package proyecto_programacion;
 import java.util.*;
 
 class administracion {
-	ArrayList<vehiculos> automovil;
-	ArrayList<personas> person;
-	ArrayList<String> admin;
-	E_S teclado= new E_S();
-	
+	protected ArrayList<vehiculos> automovil= new ArrayList<vehiculos>();
+	protected ArrayList<personas> person= new ArrayList<personas>();
+	protected ArrayList<String> admin= new ArrayList<String>();
+	protected E_S teclado= new E_S();
+
+	//Cuentas de Administrador y Cuenta de Empleado
+		private void CuentaAdmin () {
+			String usuario, password, dinerobase;
+			boolean ciclo= true;
+			int contador= 2;
+
+			if (admin.size() == 0) {
+				System.out.println("\nBuenas administrador, esta es la primera vez que");
+				System.out.println("inicia sesi\u00f3n, por favor rellene los siguientes datos");
+				System.out.println("");
+				do {
+					System.out.print("Nombre de administrador: ");
+					usuario= teclado.CadenaTexto();
+					System.out.print("Contrase\u00f1a: ");
+					password= teclado.CadenaTexto();
+					System.out.print("Establezca el dinero base: ");
+					dinerobase= teclado.CadenaTexto();
+					//Comproba que los datos no esten vacios
+					if (usuario.equals("") || password.equals("") || dinerobase.equals("") ) {
+						System.out.println("\nError: Nombre de administrador, contrase\u00f1a y/o dinero base no establecidos");
+						System.out.println("Por favor vuelva a repetir todo de nuevo\n");
+					} else {
+						//Comprueba que el dinerobase no sean letras y que no contenga espacios
+						if (dinerobase.matches("[0-9]+") || dinerobase.contains(".") || dinerobase.contains(",")) {
+							//Reemplaza la coma por un .
+							dinerobase= dinerobase.replace(',', '.');
+							double dinero= Double.parseDouble(dinerobase);
+							if (dinero >= 500000 ) {
+								ciclo= false;
+							} else {
+								System.out.println("\nError: El dinero base tiene que ser mayor o igual a 500mil euros");
+								System.out.println("Por favor vuelva a repetir todo de nuevo\n");
+							}//Fin del if
+							
+						} else {
+							System.out.println("\nError: El dinero base tiene que ser un n\u00famero");
+							System.out.println("Por favor vuelva a repetir todo de nuevo\n");
+						}//Fin del if
+					}//Fin del if
+
+				} while (ciclo);
+				admin.add(usuario); //0
+				admin.add(password); //1
+				admin.add(dinerobase); //2
+				
+				System.out.println("\n == Cuenta Administrador creada con \u00e9xito == ");
+				System.out.println("");
+				System.out.println("Por favor vuelva a seleccionar el tipo de cuenta administrador");
+				System.out.println("e inicie sesi\u00f3n con los datos que ha indicado.");
+				
+			} else {
+				System.out.println("\nPor favor indique los siguientes datos: ");
+
+				do {
+					System.out.print("\nNombre de usuario: ");
+					usuario= teclado.CadenaTexto();
+					System.out.print("Contrase\u00f1a: ");
+					password= teclado.CadenaTexto();
+					
+					//Comprueba si nombre de usuario y contrasenia coinciden con el registro.
+					if (admin.get(0).equals(usuario) && admin.get(1).equals(password)) {
+						ciclo= false;
+					} else {
+						//Contador de fallos
+						if (contador > 0) {
+							System.out.println("\nError: El usuario o la contrase\u00f1a son incorrectos");
+							contador--;
+						} else {
+							System.out.println("\n == Has fallado 3 veces, volviendo al menu principal == ");
+							ciclo= false;
+						}//Fin del if
+					}//Fin del if
+				} while (ciclo); //Fin del do
+
+				//Se ejecuta si los datos son correctos
+				while (!ciclo && contador > 0) {
+					System.out.println("\n == Bienvenido Administrador "+usuario+" == ");
+					MenuAdmin();
+				}//Fin del while
+			}
+		}
+		private void CuentaEmpleado() {
+			String usuario;
+			boolean existe= false;
+			//int contador= 2;
+			admin.remove(1);
+			admin.add(1, "verdadero");
+			
+			if (admin.get(0).equals("verdadero")) {
+				System.out.println("\nError: La cuenta administrador no ha sido establecida");
+			} else {
+				System.out.println("\n == INICIO DE SESI\u00f3N EMPLEADO == ");
+				System.out.println("");
+				System.out.print("Introduzca su DNI: ");
+				usuario= teclado.CadenaTexto();
+
+				//Comprobamos que el DNI de la persona exista
+				for (int i=0; i<person.size(); i++) {
+					if (person.get(i).getDni().equals(usuario) && person.get(i).getTipo_persona().equals("empleado")) {
+						existe= true;
+					} else {
+						existe= false;
+					}//Fin del if
+					
+				}//Fin del for i
+				
+				//Segun la verificacion anterior nos dira si existe o no
+				for (int j=0; j<person.size(); j++) {
+					if (!existe) {
+						System.out.println("\nError: No existe empleado registrado con ese DNI");
+						j= person.size();
+					} else if (person.get(j).getDni().equals(usuario)) {
+						System.out.println(" == BIENVENIDO "+person.get(j).getNombre()+" "+person.get(j).getApellidos()+" == ");
+							if (((asesor) person.get(j)).getTrabajo_asesor().equals("Asesor")) {
+								System.out.println("\n == Asesor: "+person.get(j).getNombre()+" "+person.get(j).getApellidos()+" == ");
+								MenuAsesor();
+								j= person.size();
+							} else if (((asesor) person.get(j)).getTrabajo_asesor().equals("Mecanico")) {
+								System.out.println("\n == Mecanico: "+person.get(j).getNombre()+" "+person.get(j).getApellidos()+" == ");
+								MenuMecanico();
+								j= person.size();
+							}
+					}//Fin del if
+				}//Fin del for j
+			}//Fin del if
+		}
+
+	//Menus
+		protected void MenuPrincipal () {
+			String cuenta= "";
+			boolean ciclo= true, confirmacion= false;
+			
+			do {
+				do {
+					System.out.println("");
+					System.out.println(" == CONCENSIONARIO == ");
+					System.out.println("");
+					System.out.println("Por favor, a continuaci\u00f3n elija el tipo de cuenta");
+					System.out.println("");
+					System.out.println(" [1] Administrador");
+					System.out.println(" [2] Empleado");
+					System.out.println(" [3] Salir");
+					System.out.println("");
+					System.out.print("Elija la opci\u00f3n: ");
+					cuenta= teclado.CadenaTexto();
+
+					//Llamamos metodo que compruebar si la opcion 
+					//introducida es un numero.
+					if (isNumero(cuenta)) {
+						confirmacion= true;
+					} else {
+						confirmacion= false;
+					}// Fin del if
+				} while (!confirmacion); //Fin del do
+
+				switch (cuenta) {
+				case "1":
+					//Llamamos al metodo que mostrara unas opciones si es la
+					//primera vez que inicia sesion y otra si ya ha iniciado antes.
+					CuentaAdmin();
+
+					break;
+				case "2":
+					//Llamamos al metodo que nos pedira el DNI del empleado y
+					//mostrara un menu segun la categoria del empleado (asesor o mecanico)
+					CuentaEmpleado();
+
+					break;
+				case "3":
+					//Salimos del programa
+					System.out.println(" == Hasta la pr\u00f3xima == ");
+					ciclo= false;
+
+				default:
+					System.out.println("\nError: La opci\u00f3n elegida es incorrecta");
+				
+				}//Fin del swith
+			} while (ciclo);
+		}
+		protected void MenuAdmin () {
+			String datos;
+			boolean ciclo= true, confirmacion= false;
+
+			do {
+				do {
+					System.out.println("");
+					System.out.println(" [1] Alta empleado");
+					System.out.println(" [2] Baja empleado");
+					System.out.println(" [3] Mostrar empleado");
+					System.out.println(" [4] Modificar empleado");
+					System.out.println(" [5] Alta vehiculo");
+					System.out.println(" [6] Baja vehiculo");
+					System.out.println(" [7] Mostrar vehiculos");
+					System.out.println(" [8] Modificar vehiculo");
+					System.out.println(" [9] Cerrar sesi\u00f3n");
+					System.out.println("");
+					System.out.print("Elija una opci\u00f3n: ");
+					datos= teclado.CadenaTexto();
+
+					if (isNumero(datos)) {
+						confirmacion= true;
+					} else {
+						confirmacion= false;
+					}// Fin del if
+				
+				} while (!confirmacion);
+				
+				switch (datos) {
+				case "1":
+					AltaEmpleado();
+					System.out.println(person.get(0));
+					break;
+				case "2":
+					//BajaEmpleado(person);
+					break;
+				case "3":
+					//MostrarEmpleado(person);
+					break;
+				case "4":
+					//ModificarEmpleado(automovil);
+					break;
+				case "5":
+					//AltaVehiculo(automovil);
+					break;
+				case "6":
+					//BajaVehiculo(automovil);
+					break;
+				case "7":
+					//MostrarVehiculo(automovil);
+					break;
+				case "8":
+					//ModificarVehiculo(automovil);
+					break;
+				case "9":
+					ciclo= false;
+					admin.remove(1);
+					admin.add(1, "verdadero");
+					break;
+				}
+			} while (ciclo);
+		}
+		protected void MenuAsesor () {
+			String datos;
+			boolean ciclo= true, confirmacion= false;
+
+			do {
+				do {
+					System.out.println("");
+					System.out.println(" [1] Mostrar Cliente");
+					System.out.println(" [2] Vender vehiculo");
+					System.out.println(" [3] Comprar vehiculo");
+					System.out.println(" [4] Mostrar Vehiculo");
+					System.out.println(" [5] Devolver Vehiculo");
+					System.out.println(" [6] Enviar a mecanico");
+					System.out.println(" [7] Cerrar sesi\uf003n");
+					System.out.println("");
+					System.out.print("Elija una opci\u00fan: ");
+					datos= teclado.CadenaTexto();
+					
+					if (isNumero(datos)) {
+						confirmacion= true;
+					} else {
+						confirmacion= false;
+					}// Fin del if
+					
+				} while (!confirmacion);
+				
+				switch (datos) {
+				case "1":
+					//VentaVehiculo(person, automovil);
+
+					break;
+				case "2":
+					//CompraVehiculo(person, automovil);
+					
+					break;
+				case "3":
+					//MostrarCliente(person);
+
+					break;
+				case "4":
+					//MostrarCliente(person);
+
+					break;
+				case "5":
+					ciclo= false;
+					
+					break;
+				default:
+					System.out.println("\nError: La opci\u00fan elegida es incorrecta");
+					
+					break;
+				}
+				
+			}while (ciclo);
+		}
+		protected void MenuMecanico () {
+			String datos;
+			boolean ciclo= true, confirmacion= false;
+
+			do {
+				do {
+					System.out.println("");
+					System.out.println(" [1] Mostrar Vehiculo");
+					System.out.println(" [2] Marcar vehiculo reparado");
+					System.out.println(" [3] Vehiculos en reparaci\u00f3n");
+					System.out.println(" [4] Cerrar sesi\uf003n");
+					datos= teclado.CadenaTexto();
+					
+					if (isNumero(datos)) {
+						confirmacion= true;
+					} else {
+						confirmacion= false;
+					}// Fin del if
+					
+				} while (!confirmacion);
+				
+				switch (datos) {
+				case "1":
+					//MostrarVehiculo(automovil);
+					
+					break;
+				case "2":
+					//RepararVehiculo(automovil);
+					
+					break;
+				case "3":
+					//MostrarEnReparacion(automovil);
+					
+					break;
+				case "4":
+					ciclo= false;
+					
+					break;
+				default:
+					System.out.println("\nError: La opci\u00f3n elegida es incorrecta");
+					
+					break;
+				}
+				
+			}while (ciclo);
+		}
+
 	//Altas y Bajas
 	protected void AltaEmpleado () {
 		String tipoEmpleado= "", datos= "";
@@ -19,7 +362,7 @@ class administracion {
 
 		do {
 			System.out.print("Tipo de empleado: ");
-			teclado.CadenaTexto(tipoEmpleado);
+			tipoEmpleado= teclado.CadenaTexto();
 
 			if (tipoEmpleado.equals("")) {
 				System.out.println("\nError: Por favor rellene el campo para poder continuar");
@@ -31,7 +374,7 @@ class administracion {
 				//DNI
 				do {
 					System.out.print("DNI: ");
-					teclado.CadenaTexto(datos);
+					datos= teclado.CadenaTexto();
 
 					if (datos.equals("")) {
 						System.out.println("\nError: Por favor rellene el campo DNI");
@@ -48,7 +391,7 @@ class administracion {
 				//Nombre
 				do {
 					System.out.print("Nombre: ");
-					teclado.CadenaTexto(datos);
+					datos= teclado.CadenaTexto();
 
 					if (datos.equals("")) {
 						System.out.println("\nError: Por favor rellene el campo nombre");
@@ -63,7 +406,7 @@ class administracion {
 				//Apellidos
 				do {
 					System.out.print("Apellidos: ");
-					teclado.CadenaTexto(datos);
+					datos= teclado.CadenaTexto();
 					
 					if (datos.equals("")) {
 						System.out.println("\nError: Por favor rellene el campo apellidos");
@@ -78,7 +421,7 @@ class administracion {
 				//Telefono movil
 				do {
 					System.out.print("Telefono m\u00f3vil: ");
-					teclado.CadenaTexto(datos);
+					datos= teclado.CadenaTexto();
 					
 					if (datos.equals("")) {
 						System.out.println("\nError: Por favor rellene el campo telefono m\u00f3vil");
@@ -95,11 +438,11 @@ class administracion {
 				//Fecha nacimiento
 				do {
 					System.out.print("Fecha de nacimiento (dd/mm/yyyy): ");
-					teclado.CadenaTexto(datos);
+					datos= teclado.CadenaTexto();
 
 					if (datos.equals("")) {
 						System.out.println("\nError: Por favor rellene el campo fecha de nacimiento");
-					} else if (isFechaNacimiento(datos)) {
+					} else if (isFecha(datos)) {
 						a.setFecha_nacimiento(datos);
 						m.setFecha_nacimiento(datos);
 						correcto= true;
@@ -112,8 +455,8 @@ class administracion {
 				//Cuenta bancaria
 				do {
 					System.out.print("Cuenta bancaria: ");
-					teclado.CadenaTexto(datos);
-					
+					datos= teclado.CadenaTexto();
+
 					if (datos.equals("")) {
 						System.out.println("\nError: Por favor rellene el campo cuenta bancaria");
 					} else {
@@ -121,13 +464,13 @@ class administracion {
 						m.setCuenta_bancaria(datos);
 						correcto= true;
 					}//Fin del if
-					
+
 				} while (!correcto);
 				correcto= false;
 				//Direccion
 				do {
 					System.out.print("Direcci\u00f3n: ");
-					teclado.CadenaTexto(datos);
+					datos= teclado.CadenaTexto();
 					
 					if (datos.equals("")) {
 						System.out.println("\nError: Por favor rellene el campo direccion");
@@ -142,7 +485,7 @@ class administracion {
 				//Sueldo
 				do {
 					System.out.print("Sueldo: ");
-					teclado.CadenaTexto(datos);
+					datos= teclado.CadenaTexto();
 
 					if (datos.equals("")) {
 						System.out.println("\nError: Por favor rellene el campo nombre");
@@ -271,7 +614,7 @@ class administracion {
 		System.out.println("\nPor favor, a continuaci\u00f3n indique el tipo de vehiculo");
 		System.out.println("Coche - Moto | Si desea cancelar el alta escriba Cancelar\n");
 		System.out.print("Tipo de vehiculo: ");
-		teclado.CadenaTexto(tipoVehiculo);
+		tipoVehiculo= teclado.CadenaTexto();
 
 		do {
 			if (tipoVehiculo.equals("")) {
@@ -283,7 +626,7 @@ class administracion {
 				//Matricula
 				do {
 					System.out.print("Matr\u00edcula: ");
-					teclado.CadenaTexto(datos);
+					datos= teclado.CadenaTexto();
 
 					if (datos.equals("")) {
 						System.out.println("\nError: Por favor rellene el campo matr\u00edcula");
@@ -300,7 +643,7 @@ class administracion {
 				//Modelo
 				do {
 					System.out.print("Modelo: ");
-					teclado.CadenaTexto(datos);
+					datos= teclado.CadenaTexto();
 
 					if (datos.equals("")) {
 						System.out.println("\nError: Por favor rellene el campo modelo");
@@ -315,7 +658,7 @@ class administracion {
 				//Color
 				do {
 					System.out.print("Color: ");
-					teclado.CadenaTexto(datos);
+					datos= teclado.CadenaTexto();
 					
 					if (datos.equals("")) {
 						System.out.println("\nError: Por favor rellene el campo apellidos");
@@ -330,7 +673,7 @@ class administracion {
 				//Combustible
 				do {
 					System.out.print("Combustible: ");
-					teclado.CadenaTexto(datos);
+					datos= teclado.CadenaTexto();
 					
 					if (datos.equals("")) {
 						System.out.println("\nError: Por favor rellene el campo combustible");
@@ -348,7 +691,7 @@ class administracion {
 				//plazas
 				do {
 					System.out.print("Plazas: ");
-					teclado.CadenaTexto(datos);
+					datos= teclado.CadenaTexto();
 					if (datos.equals("")) {
 						System.out.println("\nError: Por favor rellene el campo plazas");
 					} else if (isNumero(datos)) {
@@ -369,7 +712,7 @@ class administracion {
 				//Kilometros
 				do {
 					System.out.print("Kilometros: ");
-					teclado.CadenaTexto(datos);
+					datos= teclado.CadenaTexto();
 					
 					if (datos.equals("")) {
 						System.out.println("\nError: Por favor rellene el campo kilometros");
@@ -386,7 +729,7 @@ class administracion {
 				//Precio
 				do {
 					System.out.print("Precio: ");
-					teclado.CadenaTexto(datos);
+					datos= teclado.CadenaTexto();
 					
 					if (datos.equals("")) {
 						System.out.println("\nError: Por favor rellene el campo precio");
@@ -449,7 +792,7 @@ class administracion {
 	protected void RepararVehiculo () {
 		
 	}
-	
+
 	//Comprobacion de valides
     protected boolean isDNI (String dni) {  	
         boolean valido = false;
@@ -469,30 +812,63 @@ class administracion {
         }//Fin del if
         
         if(valido) {
-        	for (int j= 0; j < person.size(); j++) {
-        		if (person.get(j).getDni().equalsIgnoreCase(dni)) {
-        			System.out.println("\nError: Ya existe una persona registrada con ese DNI");
-        			j= person.size();
-        			valido= false;
-        		} else {
-        			//Validamos que la letra corresponda al DNI
-                    letra = Character.toUpperCase(dni.charAt(8));
-                    miDNI = Integer.parseInt(dni.substring(0,8));
-                    resto = miDNI % 23;
-                    valido = (letra == asigLetra[resto]);
+        	if (person.size() > 0 ) {
+        		for (int j= 0; j < person.size(); j++) {
+            		if (person.get(j).getDni().equalsIgnoreCase(dni)) {
+            			System.out.println("\nError: Ya existe una persona registrada con ese DNI");
+            			j= person.size();
+            			valido= false;
+            		} else {
+            			//Validamos que la letra corresponda al DNI
+                        letra = Character.toUpperCase(dni.charAt(8));
+                        miDNI = Integer.parseInt(dni.substring(0,8));
+                        resto = miDNI % 23;
+                        valido = (letra == asigLetra[resto]);
+            		}
         		}
+        	} else {
+        		//Validamos que la letra corresponda al DNI
+                letra = Character.toUpperCase(dni.charAt(8));
+                miDNI = Integer.parseInt(dni.substring(0,8));
+                resto = miDNI % 23;
+                valido = (letra == asigLetra[resto]);
         	}
         }//Fin del if
         
         return valido;
     }
-	protected boolean isNumero (String opcion) {
+    protected boolean isFecha (String fecha) {
+		boolean valido= false;
+		int[] diasMes= {31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31};
+
+		if (fecha.matches("^(3[01]|[12][0-9]|0[1-9])/(1[0-2]|0[1-9])/[0-9]{4}$")) {
+			int dia= Integer.parseInt(fecha.substring(0, 2)); 
+			int mes= Integer.parseInt(fecha.substring(3, 5)); 
+			int anio= Integer.parseInt(fecha.substring(6, 10)); 
+
+			if (dia > 28 && mes == 2) {
+				if (((anio%400 == 0 ) || (anio%100 != 0)) && (anio%4 == 0)) {
+					valido= true;
+				} else {
+					System.out.println("\nError: Este a\u00f1o no es bisiesto");
+					valido= false;
+				}
+			} else {
+				valido= dia > 0 && dia <= diasMes[mes - 1];
+			}
+		} else {
+			valido= false;
+		}
+
+		return valido;
+	}
+	protected boolean isNumero (String numero) {
 		boolean verdadero= false;
 		
-		if (opcion.matches("[0-9]+")) {
+		if (numero.matches("[0-9]+")) {
 			verdadero= true;
 		} else {
-			System.out.println("\nError: La opcion debe ser un numero.");
+			System.out.println("\nError: La opci\u00f3n debe ser un n\u00famero.");
 			verdadero= false;
 		}//Fin del if
 		
@@ -541,29 +917,5 @@ class administracion {
 		
 		return valido;
 	}
-	protected boolean isFechaNacimiento (String fecha) {
-		boolean valido= false;
-		int[] diasMes= {31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31};
 
-		if (fecha.matches("^(3[01]|[12][0-9]|0[1-9])/(1[0-2]|0[1-9])/[0-9]{4}$")) {
-			int dia= Integer.parseInt(fecha.substring(0, 2)); 
-			int mes= Integer.parseInt(fecha.substring(3, 5)); 
-			int anio= Integer.parseInt(fecha.substring(6, 10)); 
-
-			if (dia > 28 && mes == 2) {
-				if (((anio%400 == 0 ) || (anio%100 != 0)) && (anio%4 == 0)) {
-					valido= true;
-				} else {
-					System.out.println("\nError: Este a\u00f1o no es bisiesto");
-					valido= false;
-				}
-			} else {
-				valido= dia > 0 && dia <= diasMes[mes - 1];
-			}
-		} else {
-			valido= false;
-		}
-
-		return valido;
-	}
 }
