+++ proyecto_programacion/concesionario.java
@@ -3,7 +3,7 @@ package proyecto_programacion;
 //euro -> \u20ac a -> \u00e1 e -> \u00e9 i -> \u00ed o -> \u00f3 u -> \u00fa n -> \u00f1
 public class concesionario {
 	public static void main(String[] args) {
-		menus invocar= new menus();
+		administracion invocar= new administracion();
 
 		invocar.MenuPrincipal();
 	}//Fin del main
